+++ app/src/js/utils/table-config/execution-status.js
@@ -1,4 +1,5 @@
 /* eslint-disable react/prop-types */
+import path from 'path';
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 import get from 'lodash/get';
@@ -10,102 +11,61 @@ import DefaultModal from '../../components/Modal/modal';
 
 export const tableColumns = [
   {
-    Header: 'Id',
+    Header: 'Event ID',
     accessor: 'id',
-    width: 10
+    width: 10,
+    Cell: ({ cell: { value } }) => {
+      if (value) {
+        return (
+          <>
+            <span className="num-title--round">
+              {value}
+            </span>
+          </>
+        );
+      }
+      return 'N/A';
+    }
   },
   {
-    Header: 'Type',
-    accessor: 'type',
-    Cell: ({ cell: { value }, row: { original: { eventDetails } } }) => {
+    Header: 'Step Name',
+    accessor: 'name',
+    Cell: ({ cell: { value }, row: { original: { id, eventDetails } } }) => {
       const [showModal, setShowModal] = useState(false);
-      const { id } = eventDetails || {};
       function toggleModal(e) {
         if (e) {
           e.preventDefault();
         }
         setShowModal(!showModal);
       }
-      if (eventDetails) {
-        let buttonClass;
-        if (eventDetails.type === 'LambdaFunctionFailed') {
-          buttonClass = 'button button--small button--no-left-padding button--failed';
-        } else {
-          buttonClass = 'button button--small button--no-left-padding';
-        }
-        return (
-          <>
-            <button
-              onClick={toggleModal}
-              className={buttonClass}
-            >
-              {eventDetails.type}
-            </button>
-            <DefaultModal
-              showModal={showModal}
-              title={`ID ${id}: ${value}`}
-              onCloseModal={toggleModal}
-              hasConfirmButton={false}
-              cancelButtonClass="button--close"
-              cancelButtonText="Close"
-              className="execution__modal"
-            >
-              <pre>{JSON.stringify(eventDetails, null, 2)}</pre>
-            </DefaultModal>
-          </>
-        );
-      }
-      return 'N/A';
+      return (
+        <>
+          <span className="link link--pad-right" onClick={toggleModal} role="button" tabIndex="0">
+            {value || 'N/A'}
+          </span>
+          {eventDetails.type.toLowerCase().includes('failed')
+            ? <i className="fas fa-times-circle status-icon--failed"></i>
+            : <i className="far fa-check-circle status-icon--success"></i>}
+          <DefaultModal
+            showModal={showModal}
+            title={`ID ${id}: ${value || 'N/A'} - ${eventDetails.type}`}
+            onCloseModal={toggleModal}
+            hasConfirmButton={false}
+            cancelButtonClass="button--close"
+            cancelButtonText="Close"
+            className="execution__modal"
+          >
+            <pre>{JSON.stringify(eventDetails, null, 2)}</pre>
+          </DefaultModal>
+        </>
+      );
     },
-  },
-  {
-    Header: 'Step Name',
-    accessor: 'name',
   },
   {
     Header: 'Timestamp',
     accessor: 'timestamp',
     Cell: ({ cell: { value } }) => fullDate(value)
   },
-  {
-    Header: 'Event Details',
-    accessor: 'eventDetails',
-    Cell: ({ cell: { value } }) => {
-      const [showModal, setShowModal] = useState(false);
-      const { id } = value || {};
-      function toggleModal(e) {
-        if (e) {
-          e.preventDefault();
-        }
-        setShowModal(!showModal);
-      }
-      if (value) {
-        return (
-          <>
-            <button
-              onClick={toggleModal}
-              className="button button--small button--no-left-padding"
-            >
-              More Details
-            </button>
-            <DefaultModal
-              showModal={showModal}
-              title={`ID ${id}: Event Details`}
-              onCloseModal={toggleModal}
-              hasConfirmButton={false}
-              cancelButtonClass="button--close"
-              cancelButtonText="Close"
-              className="execution__modal"
-            >
-              <pre>{JSON.stringify(value, null, 2)}</pre>
-            </DefaultModal>
-          </>
-        );
-      }
-      return 'N/A';
-    },
-    disableSortBy: true,
-  }
 ];
 
 export const metaAccessors = ({
@@ -177,6 +137,26 @@ export const metaAccessors = ({
     },
   },
   {
+    label: 'Granule ID',
+    property: 'granules',
+    accessor: (d) => {
+      if (!d) return 'N/A';
+      return (
+        <Link to={() => ({ pathname: `/granules/granule/${encodeURIComponent(path.basename(d[0].granuleId))}` })}>{d[0].granuleId}</Link>
+      );
+    },
+  },
+  {
+    label: 'Associated Executions List',
+    property: 'granules',
+    accessor: (d) => {
+      if (!d) return 'N/A';
+      return (
+        <Link to={() => ({ pathname: `/executions/executions-list/${encodeURIComponent(d[0].collectionId)}/${encodeURIComponent(path.basename(d[0].granuleId))}` })}>Link</Link>
+      );
+    },
+  },
+  {
     label: 'Input',
     property: 'input',
     accessor: (d) => {
@@ -187,7 +167,7 @@ export const metaAccessors = ({
               onClick={() => toggleModal('input')}
               className="button button--small button--no-left-padding"
             >
-              Show Input
+                Show Input
             </button>
             <DefaultModal
               showModal={showInputModal}
@@ -212,20 +192,20 @@ export const metaAccessors = ({
     accessor: (d) => {
       if (d) {
         const jsonData =
-          typeof Blob !== 'undefined'
-            ? new Blob([d], { type: 'text/json' })
-            : null;
+            typeof Blob !== 'undefined'
+              ? new Blob([d], { type: 'text/json' })
+              : null;
         const downloadUrl =
-          typeof window.URL.createObjectURL === 'function'
-            ? window.URL.createObjectURL(jsonData)
-            : '';
+            typeof window.URL.createObjectURL === 'function'
+              ? window.URL.createObjectURL(jsonData)
+              : '';
         return (
           <>
             <button
               onClick={() => toggleModal('output')}
               className="button button--small button--no-left-padding"
             >
-              Show Output
+                Show Output
             </button>
             <DefaultModal
               showModal={showOutputModal}
@@ -238,7 +218,7 @@ export const metaAccessors = ({
                     download="output.json"
                     href={downloadUrl}
                   >
-                    Download File
+                      Download File
                   </a>
                 </>
               }
@@ -262,11 +242,11 @@ export const metaAccessors = ({
     accessor: (d) => {
       const kibanaLink = kibanaExecutionLink(cumulusInstance, d);
       const className =
-        'button button--small button__goto button__arrow button__animation button__arrow--white';
+          'button button--small button__goto button__arrow button__animation button__arrow--white';
       if (kibanaLink && kibanaLink.length) {
         return (
           <a href={kibanaLink} target="_blank" className={className}>
-            View Logs in Kibana
+              View Logs in Kibana
           </a>
         );
       }
@@ -280,7 +260,7 @@ export const metaAccessors = ({
           title={`${d}/logs`}
           className={className}
         >
-          View Execution Logs
+            View Execution Logs
         </Link>
       );
     },